--- packages/chain-ethereum-eip712/.gitignore
@@ -1,3 +1,0 @@
-artifacts
-cache
-typechain-types
