--- packages/chain-ethereum-eip712/README.md
@@ -1,18 +1,0 @@
-# @canvas-js/chain-ethereum-eip712
-
-## Table of Contents
-
-- [Installation](#installation)
-- [API](#api)
-
-## Installation
-
-```
-npm i @canvas-js/chain-ethereum-eip712
-```
-
-## API
-
-```ts
-TODO: Replace with new class definition
-```
