--- packages/chain-ethereum-eip712/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@canvas-js/chain-ethereum-eip712",
-  "version": "0.8.21",
-  "type": "module",
-  "author": "Canvas Technologies, Inc. (https://canvas.xyz)",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "lib"
-  ],
-  "imports": {
-    "#target": {
-      "node": "./lib/targets/node/index.js",
-      "browser": "./lib/targets/browser/index.js",
-      "default": "./lib/targets/default/index.js"
-    }
-  },
-  "dependencies": {
-    "@canvas-js/interfaces": "0.8.21",
-    "@canvas-js/signed-cid": "0.8.21",
-    "@ipld/dag-json": "^10.1.5",
-    "@libp2p/logger": "^4.0.2",
-    "multiformats": "^12.1.3"
-  }
-}
