--- packages/chain-ethereum-eip712/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./EIP712Signer.js"
-export { validateSessionData } from "./utils.js"
-export { eip712TypeDefinitions } from "./types.js"
