--- packages/chain-ethereum-eip712/src/targets/default/index.ts
@@ -1,10 +1,0 @@
-import type { PlatformTarget } from "../index.js"
-
-export default {
-	getSessionStore() {
-		throw new Error("unsupported platform")
-	},
-	getRandomValues() {
-		throw new Error("unsupported platform")
-	},
-} satisfies PlatformTarget
