--- packages/chain-ethereum-eip712/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "include": ["src"],
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "lib",
-    "module": "Node16",
-    "moduleResolution": "Node16"
-  },
-  "references": [
-    {"path": "../interfaces"},
-    {"path": "../signed-cid"}
-  ]
-}
