+++ packages/interfaces/test/signers.test.ts
@@ -5,9 +5,8 @@ import { verifySignedValue } from "@canvas-js/signed-cid"
 
 import { CosmosSigner } from "@canvas-js/chain-cosmos"
 import { NEARSigner } from "@canvas-js/chain-near"
-import { SIWESigner } from "@canvas-js/chain-ethereum"
+import { SIWESigner, EIP712Signer } from "@canvas-js/chain-ethereum"
 import { SIWESignerViem } from "@canvas-js/chain-ethereum-viem"
-import { EIP712Signer } from "@canvas-js/chain-ethereum-eip712"
 import { SolanaSigner } from "@canvas-js/chain-solana"
 import { SubstrateSigner } from "@canvas-js/chain-substrate"
 // import { ATPSigner } from "@canvas-js/chain-atp"
