+++ packages/interfaces/test/tsconfig.json
@@ -11,7 +11,6 @@
     { "path": "../../chain-cosmos" },
     { "path": "../../chain-ethereum" },
     { "path": "../../chain-ethereum-viem" },
-    { "path": "../../chain-ethereum-eip712" },
     { "path": "../../chain-near" },
     { "path": "../../chain-solana" },
     { "path": "../../chain-substrate" }
